if auditing is disabled, security marking param shouldn't be required
